// src/types/cellSprite.ts
export enum CellSprite {
  GROUND_A = 'ground-a',
  GROUND_B = 'ground-b',
  GROUND_C = 'ground-c',
  TREE = 'tree',
  HOUSE_BL = 'house-bl',
  HOUSE_BC = 'house-bc',
  HOUSE_BR = 'house-br',
  HOUSE_TL = 'house-tl',
  HOUSE_TC = 'house-tc',
  HOUSE_TR = 'house-tr',
  HOUSE_ML = 'house-ml',
  HOUSE_MC = 'house-mc',
  HOUSE_MR = 'house-mr',
  HOUSE_DOOR ='house-door',
}